Fixes and edits
Honed in on some of the parameter choices.  Fixed Majority vote so that it isn't using the AdaBoost model anymore.
Strange error occurring on Majority Soft Voting:
DeprecationWarning: The truth value of an empty array is ambiguous. Returning False, but in future this will result in an error. Use `array.size > 0` to check that an array is not empty.
  if diff
But apparently the approximation is very good.  I will return to this later to edit it.